feat(sgp4)!: make Error::SatRecInit carry typed SGP4Error
Move the SGP4Error enum (and its i32 conversions) from sgp4::sgp4_impl
into sgp4::error, and change Error::SatRecInit(i32) to
Error::SatRecInit(SGP4Error). The raw Vallado init error code is mapped
to the corresponding SGP4Error variant at construction. SGP4Error stays
re-exported at satkit::sgp4::SGP4Error, so the public path is unchanged.
