+++ CHANGELOG.md
@@ -35,6 +35,14 @@
 
 ### Changed
 
+- **BREAKING: `sgp4::Error::SatRecInit` now carries a typed `SGP4Error`
+  instead of a raw `i32`.** The raw Vallado init error code is mapped to the
+  corresponding `SGP4Error` variant (eccentricity, mean motion, perturbed
+  eccentricity, semi-latus rectum, orbit decay) at construction, so the
+  `Display` output is now a description rather than a bare number.
+  Additionally, the `SGP4Error` enum has moved from `sgp4::sgp4_impl` to
+  `sgp4::error`; it remains re-exported at `satkit::sgp4::SGP4Error`, so the
+  public path is unchanged.
 - **`itrfcoord(...)`, `sgp4(...)`, and `satstate.propagate(...)` now reject
   unknown keyword arguments** instead of silently ignoring them. Previously a
   typo such as `itrfcoord(..., alttiude=100)` was dropped, leaving the ground
@@ -49,6 +57,18 @@
   - `time.from_rfc3339`: `s` → `rfc3339`;  `time.from_datetime`: `tm` → `dt`
   - `time.strftime`: `fmt` → `format`;  `time.strptime`: `(s, fmt)` → `(date_string, format)`
   - `kepler.from_pv`: `(r, v)` → `(pos, vel)`
+=======
+### Changed
+
+- **BREAKING: `sgp4::Error::SatRecInit` now carries a typed `SGP4Error`
+  instead of a raw `i32`.** The raw Vallado init error code is mapped to the
+  corresponding `SGP4Error` variant (eccentricity, mean motion, perturbed
+  eccentricity, semi-latus rectum, orbit decay) at construction, so the
+  `Display` output is now a description rather than a bare number.
+  Additionally, the `SGP4Error` enum has moved from `sgp4::sgp4_impl` to
+  `sgp4::error`; it remains re-exported at `satkit::sgp4::SGP4Error`, so the
+  public path is unchanged.
+>>>>>>> c87373c (feat(sgp4)!: make Error::SatRecInit carry typed SGP4Error)
 
 
 ## 0.20.2 - 2026-07-03
