+++ src/sgp4/mod.rs
@@ -40,10 +40,9 @@ mod sgp4_impl;
 mod sgp4_lowlevel;
 mod sgp4init;
 
-pub use error::{Error, Result};
+pub use error::{Error, Result, SGP4Error};
 pub use sgp4_impl::sgp4;
 pub use sgp4_impl::sgp4_full;
-pub use sgp4_impl::SGP4Error;
 pub use sgp4_impl::SGP4State;
 
 /// Canonical inputs required to initialize an SGP4 `SatRec`.
